backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
clickhouse-driver==0.2.6
pandas==2.1.4
python-jose==3.3.0
python-multipart==0.0.6